refactor: 움직임 예외 exception 패키지 -> move 패키지로 이동
